int to double in constructor of bcfperspectivecamera

+++ Xbim.BCF/XMLNodes/BCFPerspectiveCamera.cs
@@ -91,7 +91,7 @@ namespace Xbim.BCF.XMLNodes
         private BCFPerspectiveCamera()
         { }
 
-        public BCFPerspectiveCamera(Vector cameraViewPoint, Vector cameraDirection, Vector cameraUpVector, int fieldOfView)
+        public BCFPerspectiveCamera(Vector cameraViewPoint, Vector cameraDirection, Vector cameraUpVector, double fieldOfView)
         {
             CameraViewPoint = cameraViewPoint;
             CameraDirection = cameraDirection;
